refactor plans
